worker/requirements.txt
autogluon.tabular>=1.1.1
celery>=5.3.0
redis>=5.0.0
pandas>=2.2.0
